JobService: fix initial theme dir creation
This ensures that /data/system/theme is created with proper
permissions before any jobs are run. This fixes cases where
jobs authorized under a different context are forced to create
the theme dir under lesser privileges. This removes the need of
core commits for theme dir entirely, but depends on up-to-date
interfacer sepolicy commits.

Change-Id: Id97643897dc906104429853efa6cf518ec4f3f6b